server/requirements.txt
flask==2.0.1
flask-cors==3.0.10
transformers==4.30.2
torch
python-dotenv==0.19.0
pydantic==1.10.12